ingestion/requirements.txt
boto3>=1.35.0,<2.0.0
pypdf>=5.0.0,<6.0.0
amazon-textract-textractor>=1.8.0,<2.0.0
amazon-textract-caller>=0.2.0,<1.0.0
lancedb>=0.13.0,<1.0.0
aws-lambda-powertools>=2.19.0,<3.0.0